Hide file store object and provide thin wrapper functions around its functionality
This is to support mocking out the Application struct for testing in the CLI. Initially I wanted to mock the FileStore struct directly and not have to do this, but it seems like mocking an object owned by another crate is not very easy. Probably the solution is to not be so zealous about splitting things up into different crates, so in the future it might be a good idea to reoganize things and revisit this.