Implement help command
If no specific command is given, "help" prints the list of the commands.

This could have been factored such that help is a normal command and lives next to the others. However, since help needs to know what the other commands are and it's unlikely anything else would also need this, it seemed undesireble to plumb the command list to all commands just to enable help to have access to it. So I ended making this one command be special and live in the CLI app code itself.