Fix the tests
Took a long time to get Faux to play nice with a function that returns a `&mut dyn Write`. Not sure exactly why this is the thing that ended up working, but it's good enough for now.

Also moved the creation of the TerminalInterface (rename probably coming soon) into the outermost main function. This is to prevent the tests from running it's `create` function and consequently setting  the terminal into raw mode. The added benefit of this is that most of the functionality of the main module should theoretically be testible with mocks now.